+++ BusinessObject/SalesForce/MessageContainer.cs
@@ -8,7 +8,7 @@ namespace BusinessObject.SalesForce
         public class UI
         {
             public static string CreationSuccessMessage(string entityName, string accountName) => $"{entityName} \"{accountName}\" was created.";
-            public static string DeleteSuccessMessage(string entityName, string accountName) => $"{entityName} \"{accountName}\" was deleted. Undo" ;
+            public static string DeleteSuccessMessage(string entityName, string accountName) => $"{entityName} \"{accountName}\" was deleted. Undo";
         }
 
         public class API
@@ -25,7 +25,7 @@ namespace BusinessObject.SalesForce
             public static Error ErrorRequiredFieldMissing(string propertyName) => new Error() { ErrorCode = requiredFieldMissingCode, Message = $"Required fields are missing: [{propertyName}]", Fields = new[] { propertyName } };
 
             public static Error ErrorJsonParse(string propertyValue) => new Error() { ErrorCode = parseErrorCode, Message = $"Cannot deserialize instance of date from VALUE_STRING value {propertyValue} or request may be missing a required field at" };
-            
+
             public static Error ErrorInvalidField(string propertyName, string sobjectType) => new Error() { ErrorCode = invalidFieldCode, Message = $"No such column '{propertyName}' on sobject of type {sobjectType}" };
 
         }
