+++ BusinessObject/SalesForce/UI/Steps/UiSteps.cs
@@ -1,0 +1,10 @@
+﻿
+
+namespace BusinessObject.SalesForce.UI.Steps
+{
+    public class UiSteps
+    {
+        public AccountSteps accountSteps = new();
+        public ContactSteps contactSteps = new();
+    }
+}
