+++ Core/Helpers/ReflectionHelper.cs
@@ -1,47 +1,43 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Reflection;
-using System.Text;
-using System.Threading.Tasks;
+﻿using System.Reflection;
+
 
 namespace Core.Helpers
 {
     public class ReflectionHelper
     {
         /// <summary> 
-        /// Получаем список имен свойств 
+        /// Get property names
         /// </summary> 
-        /// <param name="currentObject"></param> 
-        /// <returns></returns> 
-        public static IEnumerable<string> GetPropertyNames(object currentObject)
+        /// <param name="_object">Object</param> 
+        /// <returns>Collection of property names</returns> 
+        public static IEnumerable<string> GetPropertyNames(object _object)
         {
-            return currentObject.GetType().BaseType.GetProperties(BindingFlags.Instance | BindingFlags.NonPublic | BindingFlags.Public).Concat(
-                            currentObject.GetType().GetProperties(BindingFlags.Instance | BindingFlags.NonPublic | BindingFlags.Public)).Select(r => r.Name.ToLower());
+            return _object.GetType().BaseType.GetProperties(BindingFlags.Instance | BindingFlags.NonPublic | BindingFlags.Public).Concat(
+                            _object.GetType().GetProperties(BindingFlags.Instance | BindingFlags.NonPublic | BindingFlags.Public)).Select(r => r.Name.ToLower());
         }
 
         /// <summary> 
-        /// Получаем значение свойства объекта 
+        /// Get property value
         /// </summary> 
-        /// <param name="_object">объект</param> 
-        /// <param name="attributeName">название свойства</param> 
+        /// <param name="_object">Object</param> 
+        /// <param name="propertyName">Property name</param> 
         /// <returns></returns> 
-        public static object GetPropertyValue(string attributeName, object _object)
+        public static object GetPropertyValue(string propertyName, object _object)
         {
-            return ReflectionHelper.GetProperties(_object)
-                                    .FirstOrDefault(p => p.Name.ToLower().Equals(attributeName.ToLower()))
+            return GetProperties(_object)
+                                    .FirstOrDefault(p => p.Name.ToLower().Equals(propertyName.ToLower()))
                                     ?.GetValue(_object);
         }
 
         /// <summary> 
-        /// Получаем список свойств 
+        /// Get collection PropertyInfo
         /// </summary> 
-        /// <param name="currentObject"></param> 
-        /// <returns></returns> 
-        public static IEnumerable<PropertyInfo> GetProperties(object currentObject)
+        /// <param name="_object">Object</param> 
+        /// <returns>Collection of PropertyInfo</returns> 
+        public static IEnumerable<PropertyInfo> GetProperties(object _object)
         {
-            return currentObject.GetType().BaseType.GetProperties(BindingFlags.Instance | BindingFlags.NonPublic | BindingFlags.Public).Concat(
-                            currentObject.GetType().GetProperties(BindingFlags.Instance | BindingFlags.NonPublic | BindingFlags.Public)).ToList();
+            return _object.GetType().BaseType.GetProperties(BindingFlags.Instance | BindingFlags.NonPublic | BindingFlags.Public).Concat(
+                            _object.GetType().GetProperties(BindingFlags.Instance | BindingFlags.NonPublic | BindingFlags.Public)).ToList();
         }
     }
 }
