+++ SalesForceTests/Configs/appsettings.json
@@ -1,10 +1,10 @@
 {
   "API": {
     "BaseUrl": "https://ooomtsdi-dev-ed.develop.my.salesforce.com/services/data/v58.0/",
-    "Token": "Bearer  00DHr0000017Sbq!ARoAQPh1xZFT1JwC.vohffovOzAf4X4R7dEIP_vbTKtvO0XQ.EiKXArz028lVYecKS3cRFyFqQ9xlct2aYF.2o4Wc1tFNZJM"
+    "Token": "Bearer  00DHr0000017Sbq!ARoAQHegKgifaN2q9MkbKX0gUprpLSX6EXMKtld_9boddBxk54Our9.3QSJFLR249SQBtn6ZJPyh0D_Xvs3jemDubVf466fl"
   },
   "Browser": {
-    "Headless": "false",
+    "Headless": "true",
     "Type": "Chrome",
     "TimeOut": 10,
     "UserName": "margo.aleksandrovna@ya.ru",
